Split LFO into Fixed and Controlled, the latter using signal buffers
